test-env/vuln-python/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
requests==2.26.0
lxml==4.6.3 